Fix database refresh for project-instance and project-template
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/ac7x/Next-Progress?shareId=XXXX-XXXX-XXXX-XXXX).
